Feat(seed): Add placeholder project data to the seeding

+++ server/db/seed.ts
@@ -13,9 +13,23 @@ export default async function seed() {
   await db.execute(`TRUNCATE TABLE "project" RESTART IDENTITY CASCADE;`);
 
 
+  /**
+   * Insert placeholder data
+   */
 
+  const adminId = await db
+    .select({ id: schema.user.id })
+    .from(schema.user)
+    .where(eq(schema.user.name, 'admin'));
 
+  if (!adminId[0].id) console.log('No admin user found');
 
+  await db.insert(schema.project).values(
+    projectsPlaceholderData.map((project) => ({
+      ...project,
+      managerId: adminId[0].id
+    }))
+  );
 
   // // Clear verification table before seeding
   // await db.execute(`TRUNCATE TABLE "verification" RESTART IDENTITY CASCADE;`);
